rename and refactor sql module; move agent context to dedicated dir
