+++ src-tauri/src/connection/query.rs
@@ -4,7 +4,7 @@ use crate::{
     binding::model::{entity::Entity, response::MultipleResponse},
     dataverse::queryengine::QueryEngine,
     oauth::tokencache::TokenCache,
-    sql_fetchxml,
+    sql,
     Database,
 };
 
@@ -21,7 +21,7 @@ pub async fn parse_sql_to_fetchxml(
     _window: tauri::Window,
     sql: String,
 ) -> Result<FetchXmlPreview, String> {
-    let parsed = sql_fetchxml::sql_to_fetchxml(&sql).map_err(|e| e.to_string())?;
+    let parsed = sql::sql_to_fetchxml(&sql).map_err(|e| e.to_string())?;
 
     Ok(FetchXmlPreview {
         entity_set: parsed.entity_set,
@@ -36,7 +36,7 @@ pub async fn execute_sql(
     sql: String,
     _database: tauri::State<'_, Database>,
 ) -> Result<MultipleResponse<Entity>, String> {
-    let parsed = sql_fetchxml::sql_to_fetchxml(&sql).map_err(|e| e.to_string())?;
+    let parsed = sql::sql_to_fetchxml(&sql).map_err(|e| e.to_string())?;
 
     let token = TokenCache::get_token().await?;
     let query_engine = QueryEngine::new("https://jyb.crm.dynamics.com/", &token);
