+++ src-tauri/src/sql/errors.rs
